third_party/WebKit/LayoutTests/fast/dom/Window/window-postmessage-clone-really-deep-array-expected.txt

Tests that we abort cloning overdeep arrays.

PASS: 'postMessage(reallyDeepArray)' threw DataCloneError: An object could not be cloned.
PASS: eventData is done of type string
